MCQGeneration/pyproject.toml
[tool.poetry]
name = "seceval"
version = "0.1.0"
description = ""
authors = ["Atum <atumli@tencent.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
trafilatura = "^1.6.3"
beautifulsoup4 = "^4.12.2"
tqdm = "^4.66.1"
pyyaml = "^6.0.1"
pandas = "^2.1.3"
html2text = "^2020.1.16"
aiohttp = "^3.9.1"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
selenium = "^4.15.2"
langchain = "^0.0.350"
openai = "^1.3.9"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"
pandas = "^2.1.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
